+++ NEWS.md
@@ -2,6 +2,7 @@
 * Improvements
   * `plot_randomized_pattern()` now uses envelopes to plot randomized summary functions
   * `plot_randomized_pattern()` includes a quantum bar
+  * `plot_randomized_pattern()` now returns plot after each other
   * `calculate_engery()` can also calculate the energy for marked reconstructions
 * Bugfixes
 * New functionality